+++ src/variables/settype.mjs
@@ -1,4 +1,4 @@
-import arrval from './arrval.mjs';
+import arrval from './to_array.mjs';
 import boolval from './boolval.mjs';
 import floatval from './floatval.mjs';
 import intval from './intval.mjs';