+++ tests/variables.global.test.mjs
@@ -19,7 +19,7 @@ test('variables.global', () => {
     expect(is_symbol).not.toBeUndefined();
     expect(is_undefined).not.toBeUndefined();
     expect(is_callable).not.toBeUndefined();
-    expect(arrval).not.toBeUndefined();
+    expect(to_array).not.toBeUndefined();
     expect(boolval).not.toBeUndefined();
     expect(floatval).not.toBeUndefined();
     expect(doubleval).not.toBeUndefined();