+++ tests/variables.test.mjs
@@ -1,6 +1,6 @@
 import { expect, test } from '@jest/globals';
 import {
-    arrval,
+    to_array,
     boolval,
     doubleval,
     empty,
@@ -57,7 +57,7 @@ test('variables', () => {
     expect(is_symbol).not.toBeUndefined();
     expect(is_undefined).not.toBeUndefined();
     expect(is_callable).not.toBeUndefined();
-    expect(arrval).not.toBeUndefined();
+    expect(to_array).not.toBeUndefined();
     expect(boolval).not.toBeUndefined();
     expect(floatval).not.toBeUndefined();
     expect(doubleval).not.toBeUndefined();